refactor: toString to formatStudyBlock
