+++ src/studyBlock.ts
@@ -1,9 +1,6 @@
-import { getFormattedParagraphs } from "./paragraphFormatting";
-import { GospelStudyPluginSettings } from "./models/GospelStudyPluginSettings";
 import { UrlParserResult } from "./models/UrlParserResult";
 
 export class StudyBlock {
-    private _paragraphs: string[] | undefined;
     private _referenceLink: string | undefined;
     private _tag: string | undefined;
 
@@ -83,34 +80,5 @@ export class StudyBlock {
      */
     public get url(): string {
         return this._urlParserResult.url.toString();
-    }
-
-    /**
-     * Converts the StudyBlock object to a string representation based on the specified format.
-     * @param format - The format string used to generate the string representation.
-     * @returns The string representation of the StudyBlock object.
-     */
-    public toString(pluginSettings: GospelStudyPluginSettings): string {
-        let injectedText = pluginSettings.studyBlockFormat;
-
-        const formattedParagraphs:string[] = getFormattedParagraphs(this.paragraphElements, pluginSettings);
-
-        const propertyNames = Object.getOwnPropertyNames(Object.getPrototypeOf(this));
-        propertyNames.forEach((key: string): void => {
-            const value = this[key as keyof StudyBlock]; // Add type assertion to keyof StudyBlock
-            if (typeof (value) !== 'string') {
-                return; // skip function calls
-            }
-
-            injectedText = injectedText.replace(new RegExp(`{{${key}}}`, 'g'), String(value));
-        });
-
-        const paragraphsMatch = injectedText.match(/{{paragraphs:([^}]*)}}/);
-        if (paragraphsMatch) {
-            const paragraphsSeparator = paragraphsMatch[1];
-            injectedText = injectedText.replace(new RegExp(`{{paragraphs:${paragraphsSeparator}}}`, 'g'), formattedParagraphs?.join(paragraphsSeparator) || '');
-        }
-
-        return injectedText;
     }
 }
